requirements.txt
aiogram
python-dotenv
requests
groq
prometheus-client